+++ lib/src/events/isolate_lock_event.dart
@@ -11,12 +11,6 @@ abstract class IsolateLockEvent extends IsolateEvent
 
   factory IsolateLockEvent.release(Capability capability, String name) = 
     IsolateLockReleaseEvent;
-
-  @override
-  bool operator ==(Object other) =>
-    identical(this, other) ||
-    other is IsolateLockEvent &&
-    this.capability == other.capability && this.name == other.name;
 }
 
 class IsolateLockAcquireEvent extends IsolateLockEvent