+++ lib/src/isolate_exceptions.dart
@@ -20,22 +20,45 @@ class IsolateNoProcessorsAvailableException implements IsolateException
   String toString() => 'Available processors: $availableProcessors';
 }
 
-/// Thrown when error [StackTrace] is too big to return from isolate.
-class IsolateTooBigStacktraceException implements IsolateException 
+/// Thrown when a function is passed an unacceptable argument.
+class IsolateInvalidArgumentException
+  extends ArgumentError implements IsolateException
 {
-  final Type type;
+  IsolateInvalidArgumentException(invalidValue, [name, message]) : 
+    super.value(invalidValue, name, message);
+}
+
+/// Thrown when a function is passed an invalid isolate entry point.
+class IsolateInvalidEntryPointException extends IsolateInvalidArgumentException
+{
+  IsolateInvalidEntryPointException(invalidValue) : super(
+    invalidValue, 'entryPoint', 
+    'Invalid argument: entryPoint must take an IsolateContext (the context)'
+  );
+}
+
+/// Thrown when passed an incorrect number of isolates.
+class IsolateInvalidCountException extends IsolateInvalidArgumentException
+{
+  IsolateInvalidCountException(invalidValue) : super(
+    invalidValue, 'count','Invalid argument: count must be greater than 0'
+  );
+}
+
+/// Thrown when error [StackTrace] is too big to return from isolate.
+class IsolateTooBigStacktraceException extends Error implements IsolateException
+{
   final String message;
   final String isolateName;
   final String _stackTrace;
 
-  StackTrace get stackTrace => 
-    StackTrace.fromString(this._stackTrace);
+  @override
+  StackTrace get stackTrace => StackTrace.fromString(this._stackTrace);
 
-  IsolateTooBigStacktraceException(this.isolateName, Error error) :
-    this.type = error.runtimeType, 
+  IsolateTooBigStacktraceException(this.isolateName, error) :
     this.message = error.toString(), 
-    this._stackTrace = error.stackTrace.toString();
+    this._stackTrace = error is Error ? error.stackTrace.toString() : '';
 
   @override
-  String toString() => '$isolateName: <$type> $message\n$stackTrace';
+  String toString() => '$isolateName - $message\n';
 }