+++ lib/src/schedule/isolate_schedule_task.dart
@@ -1,60 +1,73 @@
 part of 'isolate_schedule.dart';
 
-enum ScheduleTaskStatus { awaiting, processing, completed }
+enum _ScheduleTaskStatus { awaiting, processing, completed, canceled }
 
 class IsolateScheduleTask<R>
 {
   final IsolateTask<R> object;
-  final StreamController<IsolateResult<R>> _streamController;
+  final _streamController = StreamController<IsolateResult<R>>.broadcast();
 
-  ScheduleTaskStatus _status = ScheduleTaskStatus.awaiting;
+  _ScheduleTaskStatus _status;
+  StreamSubscription<IsolateEvent> _subscription;
 
-  IsolateScheduleTask(this.object) :
-    this._streamController = StreamController<IsolateResult<R>>.broadcast();
-
-  Future<IsolateResult<R>> single() async
+  IsolateScheduleTask(this.object)
   {
-    try {
-      final result = await this.stream.first;
-      await this._streamController.done;
-      return result;
-    }
-    catch(_) { return null; }
+    this._status = _ScheduleTaskStatus.awaiting;
+
+    this._streamController.onCancel = () =>
+      this._status = _ScheduleTaskStatus.completed;
   }
-  
+
   void listen(Stream<IsolateEvent> stream)
   {
-    stream
-      ?.where((event) => event is IsolateResult<R>)
+    if (stream == null) return;
+    if (this._subscription != null) return;
+
+    this._status = _ScheduleTaskStatus.processing;
+
+    this._subscription = stream
       ?.where((event) => event.capability == object.capability)
       ?.listen(_update, onError: (_) {}, cancelOnError: true, onDone: close);
   }
-
-  void reset() => 
-    this._status = ScheduleTaskStatus.awaiting;
-
-  void lock() => 
-    this._status = ScheduleTaskStatus.processing;
-
-  Future<void> close() async
-  {
-    this._status = ScheduleTaskStatus.completed;
-    await this._streamController.close();
-  }
-
-  Future<void> get done async => await this._streamController.done;
-  Stream<IsolateResult<R>> get stream => this._streamController.stream;
-
-  TaskPriority get priority => this.object.priority;
-
-  bool get isAwaiting => this._status == ScheduleTaskStatus.awaiting;
-  bool get isCompleted => this._status == ScheduleTaskStatus.completed;
-  bool get isProcessing => this._status == ScheduleTaskStatus.processing;
 
   void _update(IsolateEvent event) 
   {
     if (this._streamController.isClosed) return;
     if (event is IsolateResult<R>) this._streamController.add(event);
   }
+
+  Future<void> cancel() async
+  {
+    if (this.isCompleted) return;
+
+    await this.close();
+    this._status = _ScheduleTaskStatus.canceled;
+  }
+   
+  Future<IsolateResult<R>> single() async
+  {
+    try {
+      final result = await this._streamController.stream.first;
+      await this._streamController.done;
+      return result;
+    }
+    catch(e) { return null; }
+  }
+
+  void close()
+  {
+    if (this._streamController.isClosed) return;
+    
+    this._subscription?.cancel();
+    this._streamController?.close();
+  }
+
+  TaskPriority get priority => this.object.priority;
+
+  Future<void> get done async => await this._streamController.done;
+  Stream<IsolateResult<R>> get stream => this._streamController.stream;
+
+  bool get isAwaiting => this._status == _ScheduleTaskStatus.awaiting;
+  bool get isCompleted => this._status == _ScheduleTaskStatus.completed;
 }