+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- Fixes bug where `IsolateSupervisor.dispose()` cannot correctly terminate isolates.
+- Deprecates `IsolateSupervisor.restart()`, use `IsolateSupervisor.reset()` instead.
+
 ## 1.0.0
 
 - Initial version
+